+++ src/dmote_keycap/batch.clj
@@ -3,8 +3,10 @@
 (ns dmote-keycap.batch
   (:require [clojure.spec.alpha :as spec]
             [clojure.string :refer [join split]]
-            [dmote-keycap.schema :as schema]
-            [dmote-keycap.models :as models]))
+            [dmote-keycap.data :refer [face-keys]]
+            [dmote-keycap.misc :refer [deep-merge]]
+            [dmote-keycap.models :as models]
+            [dmote-keycap.schema :as schema]))
 
 ;;;;;;;;;;;;;;
 ;; Internal ;;
@@ -20,12 +22,12 @@
   (let [entry (spec/conform ::schema/batch-entry raw)]
     (case (first entry)
       :short-form (compile-key-options options (assoc-in {} default-face-path raw))
-      :long-form (merge-with merge options raw))))
+      :long-form (deep-merge options raw))))
 
 (defn- section
   "Define scad-app assets for one section of a batch."
   [cli-options [section-options entries]]
-  (map (partial compile-key-options (merge-with merge section-options cli-options))
+  (map (partial compile-key-options (deep-merge section-options cli-options))
        entries))
 
 (defn- name-from-faces
@@ -33,11 +35,13 @@
   Filenames etc. in legend strings get no special treatment. This could change
   with a decent slugifier in a future version."
   [options]
-  (->> (get-in options [:legend :faces] {})
-    (vals)
-    (mapcat (juxt :char :importable :unimportable))
-    (remove nil?)
-    (join "-")))
+  (let [face-index (fn [[face-key _]] (.indexOf face-keys face-key))]
+    (->> (get-in options [:legend :faces] {})
+      (sort-by face-index)
+      (map second)
+      (mapcat (juxt :char :importable :unimportable))
+      (remove nil?)
+      (join "-"))))
 
 (defn- name-fn
   "Return a function to gives each asset a unique name."
@@ -46,19 +50,25 @@
     (format (str "%s-%0" n-places "d-%s")
             filename index (name-from-faces options))))
 
+(defn- define-asset
+  "Define one scad-app asset.
+  Make sure each SCAD output will have a unique name, and matching names for
+  each of its SVG files for legends."
+  [namer i entry]
+  (let [filename (namer i entry)]
+    {:name filename
+     :model-main (models/keycap (assoc entry :filename filename))
+     :minimum-face-size (:face-size entry)}))
+
 
 ;;;;;;;;;;;;;;;
 ;; Interface ;;
 ;;;;;;;;;;;;;;;
 
 (defn batch-assets
-  "Define scad-app assets from the contents of an EDN file."
+  "Define scad-app assets from the parsed contents of an EDN file."
   [cli-options sections]
   (let [entries (remove nil? (mapcat (partial section cli-options) sections))
         namer (name-fn (-> entries count str count))]
-    (map-indexed (fn [i entry]
-                   {:name (namer i entry)
-                    :model-main (models/keycap entry)
-                    :minimum-face-size (:face-size entry)})
-                 entries)))
+    (map-indexed (partial define-asset namer) entries)))